BitMasks

Documentation for BitMasks.

BitMasks.enabled_flags

BitMasks.@bitmask

BitMasks.enabled_flags — Method

Return the bitmask flags present in mask; that is, all the BitMask flags for which in(flag, mask).

BitMasks.@bitmask — Macro

@bitmask BitFlags::UInt32 begin
    FLAG_A = 1
    FLAG_B = 2
    FLAG_C = 4
end

Enumeration of bitmask flags that can be combined with &, | and xor, forbidding the combination of flags from different BitMasks.